requirements.txt
azure-mgmt-resource
azure-mgmt-datafactory
azure-identity
python-dotenv